added backend methods, moved all submission related methods of the "web" package to the submissions.go
